src/gatt-client: Introduce btd_gatt_client.
This patch introduces struct btd_gatt_client, which intends to
encapsulate the GATT database of a btd_device. btd_gatt_client tracks
attributes of a remote device and exposes them to external applications
via the D-Bus API defined in doc/gatt.

The ultimate goal is to have a single instance of btd_gatt_client for
each btd_device and allow profiles to access attributes through it,
instead of using attio. This is the first of several stages in removing
GAttrib from bluetoothd.