src/gatt-client: Implement "Characteristics" and "Descriptors" properties.
This patch implements the org.bluez.GattCharacteristic1.Descriptors and
org.bluez.GattService1.Characteristics properties. These properties return an
array of object paths that belong to the characteristics of a service and the
descriptors of a characteristic. The properties initially return an empty array
but get updated once after the corresponding characteristic or descriptor
discovery procedure has ended.

This is used to notify the client application that all characteristics and
descriptors of a service have been discovered and the service is ready to
interact with.