core: gatt: Don't export claimed services
GattService1 objects for services which are claimed by an internal
profile are no longer exported. This is to prevent conflicts that
may arise between internal plugins and external applications that
wish to access the same GATT service.

This is at least until there is a better system for managing external
claims on GATT services and all remote profiles have been switched to
shared/gatt, which prevents some of the conflicts that may arise from
writes to CCC descriptors.